narcan R package for working with multiple cause of death data

R package for working with multiple cause of death data

See also

Author

Maintainer: Mathew Kiang mathew.kiang@gmail.com

Authors:

Monica Alexander monicaalexander@berkeley.edu